# swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ResponseData": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean"
					},
					"result": {},
					"msgTitle": {
						"type": "string"
					},
					"request": {}
				},
				"required": [
					"status",
					"success",
					"result",
					"msgTitle",
					"request"
				],
				"type": "object",
				"additionalProperties": false
			},
			"createDharmDto": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"createJaatiDto": {
				"properties": {
					"dharmId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"dharmId",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"mobileNumber": {
				"properties": {
					"mobile": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"mobile"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"email": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"firstName",
					"lastName",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			}
		}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/": {
			"get": {
				"operationId": "AdminHome",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "accept-language",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/dharm": {
			"post": {
				"operationId": "CreateDharm",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData"
								}
							}
						}
					}
				},
				"tags": [
					"Dharm API"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "accept-language",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/createDharmDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData"
								}
							}
						}
					}
				},
				"tags": [
					"Dharm API"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "accept-language",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/jaati": {
			"post": {
				"operationId": "CreateJaati",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData"
								}
							}
						}
					}
				},
				"tags": [
					"Jaati API"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "accept-language",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/createJaatiDto"
							}
						}
					}
				}
			}
		},
		"/api/otp": {
			"post": {
				"operationId": "CreateOtp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData"
								}
							}
						}
					}
				},
				"tags": [
					"OTP api"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "accept-language",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/mobileNumber"
							}
						}
					}
				}
			}
		},
		"/api/users": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ResponseData"
										},
										{
											"properties": {
												"msg": {
													"type": "string"
												},
												"success": {
													"type": "boolean"
												}
											},
											"required": [
												"msg",
												"success"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"description": "Retrieves the details of an existing user.\nSupply the unique user ID from either and receive corresponding user details.",
				"tags": [
					"Users API"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData"
								}
							}
						}
					}
				},
				"tags": [
					"Users API"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "accept-language",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}